processing for water portals but portals broke

+++ RayLua/Core/mapcore.c
@@ -8,7 +8,7 @@
 #define STANDALONE 1
 #define OOS_CHECK 1
 
-int portaln=0;
+uint16_t portaln=0;
 portal portals[100] ={};
 long gnumtiles, gmaltiles, gtilehashead[1024];
 char curmappath[MAX_PATH+1]="";
